requirements.txt
flask==2.3.3
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
gensim==4.3.1
wordcloud==1.9.2
matplotlib==3.7.2
seaborn==0.12.2
google-generativeai==0.3.2 